// src/models/IoK8sApiFlowcontrolV1beta3FlowSchemaList/types.ts
import { IoK8sApiFlowcontrolV1beta3FlowSchemaModelOptions as IoK8sApiFlowcontrolV1beta3FlowSchema } from '../IoK8sApiFlowcontrolV1beta3FlowSchema';
import { IoK8sApimachineryPkgApisMetaV1ListMetaModelOptions as IoK8sApimachineryPkgApisMetaV1ListMeta } from '../IoK8sApimachineryPkgApisMetaV1ListMeta';

export interface IoK8sApiFlowcontrolV1beta3FlowSchemaListModelOptions {
  apiVersion?: string;
  items: Array<IoK8sApiFlowcontrolV1beta3FlowSchema>;
  kind?: string;
  metadata?: IoK8sApimachineryPkgApisMetaV1ListMeta;
}